requirements.txt
django
psycopg2-binary
pymongo
python-dotenv
azure-ai-formrecognizer
azure-ai-inference
azure-core
azure-communication-email
azure-identity
azure-cognitiveservices-speech
numpy
scipy